agentic-pm/requirements.txt
# Agentic PM - Multi-agent orchestration using Linear DB
claude-agent-sdk>=0.1.0
httpx>=0.27.0
pydantic>=2.0.0
loguru>=0.7.0
python-dotenv>=1.0.0
